Advanced Distributed-Parallel Training/ParameterServerTrainingOnlySusy/colab-requirements.txt
tqdm
torch==2.0.1
torchdata
torchvision
mlflow[extras]==2.2.2
torchmetrics
matplotlib
ipython[notebook]
pandas==1.5.3
kafka-python==2.0.2
jupyter
pytest
distributed
pyparsing
ipykernel 
pytest-timeout
more-itertools
python-dotenv
scikit-learn
ipyparallel
flake8
